% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRep.R, R/package-DDModeling.R
\name{Compare,DDRep,DDRep-method}
\alias{Compare,DDRep,DDRep-method}
\alias{Compare}
\alias{Compare,DDRep,character-method}
\alias{Compare-methods}
\title{Custom implementations of the Compare method for DDModeling}
\usage{
\S4method{Compare}{DDRep,DDRep}(e1, e2)

\S4method{Compare}{DDRep,character}(e1, e2)
}
\arguments{
\item{e1}{a DDRep object}

\item{e2}{either another DDRep object or a character specifying a grid_path}
}
\description{
Custom implementations of the Compare method for DDModeling
}
