% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDFit.R
\docType{class}
\name{DDFit-class}
\alias{DDFit-class}
\title{\code{DDFit} class definition}
\description{
\code{DDFit} class definition
}
\section{Slots}{

\describe{
\item{\code{INP_REP}}{\code{DDRep} object containing the to be fitted data}

\item{\code{FIT_REP}}{\code{DDRep} object containing the fitted data}

\item{\code{MODEL}}{\code{DDModel} object containing the model that was used in the fit}

\item{\code{FIT}}{\code{DDFitPar} object containing information regarding the fit}
}}

