% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDFitPar.R
\docType{class}
\name{DDFitPar-class}
\alias{DDFitPar-class}
\title{An S4 class to represent a drift diffusion model}
\description{
An S4 class to represent a drift diffusion model
}
\section{Slots}{

\describe{
\item{\code{INP_P}}{\code{data.frame} containing the fitted parameter}

\item{\code{FIT_P}}{\code{numeric} representing the value of the Fit}

\item{\code{FIT_V}}{\code{numeric} representing the value of the Fit}

\item{\code{FIT_N}}{\code{numeric} representing the number of evaluation points used in the fit}
}}

