% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDModel.R
\docType{class}
\name{DDModel-class}
\alias{DDModel-class}
\title{\code{DDModel} class definition}
\description{
\code{DDModel} class definition
}
\section{Slots}{

\describe{
\item{\code{ID}}{\code{character} that represents the name of the model to be used (i.e. "DSTP","DMC","SSP")}

\item{\code{MM}}{\code{list} of \code{matrix} that contain values which map custom parameters to correspondend modelparameters}

\item{\code{DM}}{\code{matrix} that contains the domain of all custom parameters (and grid size steps)}

\item{\code{SP}}{\code{matrix} that contains a set of simulation-parameters important for simulation}

\item{\code{RF}}{\code{list} of numeric vectors that contain the percentiles of the representation.}
}}

