% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDModel.R
\name{DDModel}
\alias{DDModel}
\title{Function to generate a DDModel S4 class object}
\usage{
DDModel(model = NULL, task = NULL, conditions = NULL,
  parameter = NULL, dt = NULL, sigma = NULL, CDF_perc = NULL,
  CAF_perc = NULL)
}
\arguments{
\item{model}{\code{character} of the name of the Model to be used (legitamite choices are "DSTP","DMC","SSP")}

\item{task}{\code{character} specifying a specific predefined modelstructure ("flanker")}

\item{conditions}{\code{character} vector of the names of conditions}

\item{parameter}{\code{character} vector of the names of custom parameters}

\item{dt}{\code{numeric} representing the integration constant of the diffusion process}

\item{sigma}{\code{numeric} representing the diffusion constant of the diffusion process}

\item{CDF_perc}{\code{Numeric} vector specifying the CDF percentiles (note: numbers equal to absolut percentiles!)}

\item{CAF_perc}{\code{Numeric}  vector specifying the CAF percentiles (note: numbers equal to boarders of segments!)}
}
\value{
\code{DDMODEL} object
}
\description{
Function to generate a DDModel S4 class object
}
