% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRep.R
\docType{class}
\name{DDRep-class}
\alias{DDRep-class}
\alias{show,DDRep-method}
\alias{plot,DDRep,ANY-method}
\alias{plot,DDRep-method}
\alias{Compare,DDRep,DDRep-method}
\alias{Compare,DDRep-method}
\alias{Compare,DDRep,character-method}
\title{\code{DDRep} class definition}
\usage{
\S4method{show}{DDRep}(object)

\S4method{plot}{DDRep,ANY}(x)

\S4method{Compare}{DDRep,DDRep}(e1, e2)

\S4method{Compare}{DDRep,character}(e1, e2)
}
\description{
\code{DDRep} class definition
}
\section{Slots}{

\describe{
\item{\code{RAW}}{\code{list} of data.frames that contain the RAW data (i.e. 3 coloumns: $cond $resp $time)}

\item{\code{REP}}{\code{list} of data.frames that contain data representations (CDF and CAF)}

\item{\code{RF}}{\code{list} of numeric vectors that contain the percentiles of the representation.}

\item{\code{PAR}}{\code{data.frame} containing the parameters associated with the representation}
}}

