% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLANKER_DATA.R
\docType{data}
\name{FLANKER_DATA}
\alias{FLANKER_DATA}
\title{Flanker-task data}
\format{A list with 64 \code{data.frames} conteining 3 variables
\describe{
  \item{cond}{flanker condition [Cong==congurent, Incong==incongruent]}
  \item{time}{reaction time of a trial}
  \item{resp}{response for a trial [1==correct,=00incorrect]}
}}
\source{
The data is a summation of multiple classical flanker task studies undergone
at the university of Konstanz.
}
\usage{
FLANKER_DATA
}
\description{
A dataset containing the reactiontimes, flanker condition and response
for 64 subjects
}
\keyword{datasets}
