% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_DDRep.R
\name{Get_DDRep}
\alias{Get_DDRep}
\title{Function to generate a DDRep from a RAW data format}
\usage{
Get_DDRep(model = NULL, data = NULL)
}
\arguments{
\item{data}{\code{string} or \code{list} of strings to a document containing the RAW data}
}
\value{
\code{list} of \coe{DDRep} object(s)
}
\description{
Function to generate a DDRep from a RAW data format
}
