% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Grid.R
\name{GetGrid}
\alias{GetGrid}
\alias{Get_Grid}
\title{Function to fit  a given DDRep to a given DDModel}
\usage{
Get_Grid(model = NULL, wd = NULL)
}
\description{
Function to fit  a given DDRep to a given DDModel
}
