% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRep.R, R/DDFit.R, R/package-DDModeling.R
\name{plot,DDRep,ANY-method}
\alias{plot,DDRep,ANY-method}
\alias{plot}
\alias{plot,DDFit,ANY-method}
\alias{plot-methods}
\title{Custom implementations of the plot method for DDModeling}
\usage{
\S4method{plot}{DDRep,ANY}(x)

\S4method{plot}{DDFit,ANY}(x)
}
\arguments{
\item{x}{to be plotted (see usage for data types)}
}
\description{
Custom implementations of the plot method for DDModeling
}
