% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{Scale_DL_Data}
\alias{Scale_DL_Data}
\title{Retrieves scale Information from your data!}
\usage{
Scale_DL_Data(data = NULL)
}
\arguments{
\item{data}{\code{List} containing INPUT and OUTPUT element (as returned from Import_GRID)}
}
\value{
DL_scale (\code{List}) for the given data
}
\description{
Retrieves scale Information from your data!
}
