% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDModel.R, R/DDRep.R, R/package-DDModeling.R
\name{show,DDModel-method}
\alias{show,DDModel-method}
\alias{show,DDRep-method}
\alias{show}
\alias{show-methods}
\title{Custom implementations of the show method for DDModeling}
\usage{
\S4method{show}{DDModel}(object)

\S4method{show}{DDRep}(object)
}
\arguments{
\item{object}{object to be shown (see usage for data types)}
}
\description{
Custom implementations of the show method for DDModeling
}
