% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDFit.R, R/package-DDModeling.R
\name{summary,DDFit-method}
\alias{summary,DDFit-method}
\alias{summary}
\alias{summary-methods}
\title{Custom implementations of the summary method for DDModeling}
\usage{
\S4method{summary}{DDFit}(object)
}
\arguments{
\item{object}{object to be summarized (see usage for data types)}
}
\description{
Custom implementations of the summary method for DDModeling
}
