% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDFit.R
\docType{class}
\name{DDFit-class}
\alias{DDFit-class}
\title{An S4 class to represent a Fit of a given DDRep to a model}
\description{
An S4 class to represent a Fit of a given DDRep to a model
}
\section{Slots}{

\describe{
\item{\code{Input_Rep}}{\code{DDRep} object containing the to be fitted data}

\item{\code{Fit_Rep}}{\code{DDRep} object containing the fitted data}

\item{\code{Model}}{\code{DDModel} object containing the model that was used in the fit}

\item{\code{Fit}}{\code{DDFitPar} object containing information regarding the fit}
}}

