% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDFitPar.R
\docType{class}
\name{DDFitPar-class}
\alias{DDFitPar-class}
\title{\code{DDFitPar} class definition}
\description{
\code{DDFitPar} class definition
}
\section{Slots}{

\describe{
\item{\code{FIT_V}}{\code{numeric} representing the value of the Fit}

\item{\code{FIT_N}}{\code{numeric} representing the number of evaluation points used in the fit}

\item{\code{START_VALUE}}{\code{character} specifying the method used for the start value problem}

\item{\code{METHOD}}{\code{character} specifying the method used for fitting}

\item{\code{S_SAMPLING}}{\code{logical} if super sampling was used}

\item{\code{TRIALS_TOAL}}{\code{numeric} specifying the total numbers of trials that were simulated}

\item{\code{TRIALS_SAMPLE}}{\code{numeric} specifying the number of trials inside a given sample}
}}

