% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-DDModeling.R
\docType{package}
\name{DDModeling}
\alias{DDModeling}
\title{DDModeling}
\description{
DDModeling pakage
}
\details{
A pakage for the easy integration of dirft diffusion models in cognitive psychology
}
\author{
Thomas Pelzer
}
