% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRep.R
\docType{class}
\name{DDRep-class}
\alias{DDRep-class}
\title{\code{DDRep} class definition}
\description{
\code{DDRep} class definition
}
\section{Slots}{

\describe{
\item{\code{RAW}}{\code{list} of data.frames that contain the RAW data (i.e. 3 coloumns: $cond $resp $time)}

\item{\code{REP}}{\code{list} of data.frames that contain data representations (CDF and CAF)}

\item{\code{RF}}{\code{list} of numeric vectors that contain the percentiles of the representation.}

\item{\code{PAR}}{\code{data.frame} containing the parameters associated with the representation}
}}

