% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRep.R
\docType{class}
\name{DDRep-class}
\alias{DDRep-class}
\title{An S4 class to represent a representation of a drift diffusion simulation}
\description{
An S4 class to represent a representation of a drift diffusion simulation
}
\section{Slots}{

\describe{
\item{\code{RAW}}{\code{list} of data.frames that contain the RAW data (i.e. 3 coloumns: $cond $Resp $time)}

\item{\code{REP}}{\code{list} of data.frames that contain data representations (CDF and CAF)}

\item{\code{RF}}{\code{list} of numeric vectors that contain the percentiles of the representation.}
}}

