% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_DDModel.R
\name{Fit_DDModel}
\alias{Fit_DDModel}
\title{Function to fit  a given DDRep to a given DDModel}
\usage{
Fit_DDModel(model = NULL, rep = NULL, grid_path = NULL)
}
\arguments{
\item{model}{\code{DDModel} object}

\item{rep}{\code{DDRep} object}

\item{grid_path}{\code{path} to a direcotey containing a .GRID fileset. If NULL the model will be fitted using 20 randomly drawn startparametersets from the model-DOMAIN.}
}
\value{
\code{DDFit} object
}
\description{
Function to fit  a given DDRep to a given DDModel
}
