% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRID.R
\name{Generate_GRID}
\alias{Generate_GRID}
\title{Generates a Grid from a given DDModel}
\usage{
Generate_GRID(model = NULL, path = NULL, name = NULL)
}
\arguments{
\item{model}{\code{DDModel} object}

\item{path}{\code{character} that specifies the full path to the directory in which the Grid should be saved}

\item{name}{\code{character} that represents the name (and subdirectory in path) of the Grid}
}
\value{
No direct return value inside of the R-session. The calculated Grid will be saved in the specified path!
}
\description{
After calling the function the user will be instructed to enter the step sizes corresponding to the parameters listet in the used model.
              Step size should allways be of an integer value, as they represent the number of evaluation points per parameter that are used in the grid.
              Note that in the given function the evaluation points are allways equally spaced concerning the corresponding parameter domain in the used model.
              Therefor, If one would like to specify the used evaluation points it is advised to specify the domain in the model.
}
