% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRID.R
\name{Import_GRID}
\alias{Import_GRID}
\title{Imports a GRID into the R environment}
\usage{
Import_GRID(grid_path = NULL, to = "frame")
}
\arguments{
\item{grid_path}{\code{path} to a directory containing a .GRID fileset. If NULL the model will be fitted using 20 randomly drawn startparametersets from the model-DOMAIN.}

\item{to}{\code{character} that specifies the format to import to (choose from "frame", "keras_data" and "DDRep")}
}
\value{
\code{list} of \code{DDRep}, \code{list} of \code{matrix} or \code{data.frame} dependend on the "to" parameter
}
\description{
Imports a GRID into the R environment
}
