% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_DDModel.R
\name{Sim_DDModel}
\alias{Sim_DDModel}
\title{Function to simulate a given DDModel by initializing with random parameters}
\usage{
Sim_DDModel(model = NULL, trials = NULL)
}
\arguments{
\item{model}{\code{DDModel} Object}

\item{trials}{\code{Numeric} specifying the number of trials per condition}
}
\value{
\code{DDRep} object
}
\description{
Function to simulate a given DDModel by initializing with random parameters
}
